Divide pytorch_functions.py into subfiles
